requirements.txt
streamlit
langchain
langchain_groq
langchain-community
psycopg2-binary
supabase
python-dotenv
transformers
